requirements.txt
pyodbc
faker
colorama
python-dotenv
boto3==1.34.0
pyarrow==13.0.0
confluent-kafka==2.3.0